refactor: 重构软件工具包架构
