+++ src/renderer/utils/logger/index.ts
@@ -19,7 +19,7 @@
  */
 
 /**
- * @file logger.ts
+ * @file index.ts
  * @description 日志工具，同时输出到控制台和文件
  * @author 鸡哥
  */
