+++ src/renderer/utils/theme/index.ts
@@ -19,7 +19,7 @@
  */
 
 /**
- * @file theme.ts
+ * @file index.ts
  * @description 主题管理工具：初始化、切换、监听系统主题变更
  * @author 鸡哥
  */
