+++ src/renderer/utils/weatherText/index.ts
@@ -19,7 +19,7 @@
  */
 
 /**
- * @file weatherText.ts
+ * @file index.ts
  * @description 天气描述文案工具：将天气描述映射为短文本，并支持 i18n 翻译回退
  * @author 鸡哥
  */
